#10131: migrate reshard
