+++ ttnn/cpp/ttnn/operations/data_movement/data_movement_pybind.hpp
@@ -30,6 +30,7 @@
 #include "ttnn/operations/data_movement/reshape/reshape_pybind.hpp"
 #include "ttnn/cpp/ttnn/operations/data_movement/sharded/interleaved_to_sharded/interleaved_to_sharded_pybind.hpp"
 #include "ttnn/cpp/ttnn/operations/data_movement/sharded/sharded_to_interleaved/sharded_to_interleaved_pybind.hpp"
+#include "ttnn/cpp/ttnn/operations/data_movement/sharded/reshard/reshard_pybind.hpp"
 #include "ttnn/operations/data_movement/indexed_fill/indexed_fill_pybind.hpp"
 #include "ttnn/cpp/ttnn/operations/data_movement/copy/copy_pybind.hpp"
 #include "ttnn/cpp/ttnn/operations/data_movement/move/move_pybind.hpp"
@@ -69,6 +70,8 @@ void py_module(py::module& module) {
     detail::py_bind_move(module);
     py_bind_sharded_to_interleaved(module);
     py_bind_interleaved_to_sharded(module);
+    py_bind_reshard(module);
+
 }
 
 }  // namespace data_movement
