+++ ttnn/cpp/ttnn/operations/data_movement/sharded/common.cpp
@@ -7,7 +7,7 @@
 #include "tt_metal/common/constants.hpp"
 
 
-namespace ttnn {
+namespace ttnn::operations::data_movement::detail {
 
 // Utility function
 uint32_t calculate_starting_idx_h(const Tensor& tensor, uint32_t num_slices, uint32_t slice_index) {
