+++ ttnn/cpp/ttnn/operations/data_movement/sharded/interleaved_to_sharded/device/interleaved_to_sharded_op.hpp
@@ -13,6 +13,7 @@ namespace ttnn::operations::data_movement {
 
 struct InterleavedToShardedDeviceOperation {
     const MemoryConfig output_mem_config;
+    const DataType output_dtype;
 
     void validate(const std::vector<Tensor> &input_tensors) const;
     std::vector<tt::tt_metal::Shape> compute_output_shapes(const std::vector<Tensor> &input_tensors) const;
