+++ ttnn/cpp/ttnn/operations/data_movement/sharded/interleaved_to_sharded/interleaved_to_sharded.hpp
@@ -15,7 +15,8 @@ struct InterleavedToShardedOperation {
     static ttnn::Tensor invoke(
         uint8_t queue_id,
         const ttnn::Tensor& input_tensor,
-        const MemoryConfig& sharded_memory_config
+        const MemoryConfig& sharded_memory_config,
+        const std::optional<DataType> & data_type_arg
         );
     static ttnn::Tensor invoke(
         uint8_t queue_id,
@@ -23,7 +24,8 @@ struct InterleavedToShardedOperation {
         const std::variant<CoreCoord, CoreRangeSet> grid,
         const std::array<uint32_t, 2> shard_shape,
         const TensorMemoryLayout shard_scheme,
-        const ShardOrientation shard_orientation
+        const ShardOrientation shard_orientation,
+        const std::optional<DataType> & data_type_arg
         );
 
 };
@@ -31,5 +33,5 @@ struct InterleavedToShardedOperation {
 
 }  // namespace operations::data_movement
 
-constexpr auto interleaved_to_sharded = ttnn::register_operation<"ttnn::interleaved_to_sharded", ttnn::operations::data_movement::InterleavedToShardedOperation>();
+constexpr auto interleaved_to_sharded = ttnn::register_operation_with_auto_launch_op<"ttnn::interleaved_to_sharded", ttnn::operations::data_movement::InterleavedToShardedOperation>();
 }  // namespace ttnn
