+++ ttnn/cpp/ttnn/operations/data_movement/sharded/reshard/device/reshard_op.hpp
@@ -1,0 +1,30 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <optional>
+
+#include "ttnn/tensor/tensor.hpp"
+#include "ttnn/run_operation.hpp"
+#include "ttnn/deprecated/tt_dnn/op_library/work_split.hpp"
+#include "tt_metal/host_api.hpp"
+
+
+namespace ttnn::operations::data_movement {
+
+struct ReshardDeviceOperation {
+    const MemoryConfig output_mem_config;
+
+    void validate_with_output_tensors(const std::vector<Tensor> &input_tensors, const std::vector<std::optional<Tensor>> &output_tensors) const;
+    std::vector<tt::tt_metal::Shape> compute_output_shapes(const std::vector<Tensor> &input_tensors) const;
+    std::vector<Tensor> create_output_tensors(const std::vector<Tensor> &input_tensors, const std::vector<std::optional<Tensor>> &output_tensors) const;
+    operation::ProgramWithCallbacks create_program(
+        const std::vector<Tensor> &input_tensors, std::vector<Tensor> &output_tensors) const;
+
+    static constexpr auto attribute_names = std::make_tuple("output_mem_config");
+    const auto attribute_values() const { return std::make_tuple(std::cref(this->output_mem_config)); }
+};
+
+}  // namespace ttnn::operations::data_movement
